update_tests

+++ src/main/java/com/acme/meetyourroommate/service/StudentServiceImpl.java
@@ -1,8 +1,8 @@
 package com.acme.meetyourroommate.service;
 
 import com.acme.meetyourroommate.domain.model.Student;
-import com.acme.meetyourroommate.domain.model.Team;
 import com.acme.meetyourroommate.domain.repository.StudentRepository;
+import com.acme.meetyourroommate.domain.model.Team;
 import com.acme.meetyourroommate.domain.repository.TeamRepository;
 import com.acme.meetyourroommate.domain.service.StudentService;
 import com.acme.meetyourroommate.exception.ResourceNotFoundException;
@@ -11,7 +11,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 
 @Service
@@ -56,7 +55,7 @@ public class StudentServiceImpl implements StudentService {
         student.setAddress(studentRequest.getAddress());
         student.setDescription(studentRequest.getDescription());
         student.setHobbies(studentRequest.getHobbies());
-        student.setbSmoker(studentRequest.getbSmoker());
+        student.setSmoker(studentRequest.getSmoker());
 
         return studentRepository.save(student);
     }
@@ -75,12 +74,13 @@ public class StudentServiceImpl implements StudentService {
     }
 
     @Override
-    public Student joinTeam(Team team, Long studentId, Pageable pageable) {
+    public Student joinTeam(Team team, Long studentId) {
         Student student = getStudentById(studentId);
         if(student.getTeam() != null)
             return student;
 
         Team existingTeam = teamRepository.findByName(team.getName()).orElse(null);
+
         if(existingTeam==null){
             student.setTeam(team);
             return studentRepository.save(student);
@@ -91,17 +91,20 @@ public class StudentServiceImpl implements StudentService {
 
 
     @Override
-    public ResponseEntity<?> leaveTeam(Long studentId, Pageable pageable) {
+    public Student leaveTeam(Long studentId, Pageable pageable) {
         Student student = getStudentById(studentId);
         Team team = student.getTeam();
         student.setTeam(null);
 
         studentRepository.save(student);
+        if(student.getTeam() == null)
+            return student;
+
         List<Student> students = getAllStudentsByTeamId(team.getId(),pageable).getContent();
         if(students.size() == 0)
             teamRepository.delete(team);
 
-        return ResponseEntity.ok().build();
+        return student;
     }
 
 }
